Moved CourseChoicesRowHelper and CourseFeeRowHelper
